+++ lab/steps.py
@@ -72,7 +72,12 @@ class Step(object):
                                ', '.join(['%s=%s' % (k, repr(v))
                                           for (k, v) in self.kwargs.items()]))
 
+    _predefined_steps_warning = (
+        'Predefined steps have been deprecated in version 1.8. '
+        'Please define your extra experiment steps manually.')
+
     @classmethod
+    @tools.deprecated(_predefined_steps_warning)
     def zip_exp_dir(cls, exp):
         """
         Return a Step that creates a compressed tarball containing the
@@ -81,12 +86,17 @@ class Step(object):
 
             exp.add_step(Step.zip_exp_dir(exp))
 
+        .. deprecated:: 1.8
+
+            Please define your extra experiment steps manually.
+
         """
         return cls('zip-exp-dir', call,
                    ['tar', '-cjf', exp.name + '.tar.bz2', exp.name],
                    cwd=os.path.dirname(exp.path))
 
     @classmethod
+    @tools.deprecated(_predefined_steps_warning)
     def unzip_exp_dir(cls, exp):
         """
         Return a Step that unzips a compressed tarball containing the
@@ -94,18 +104,27 @@ class Step(object):
 
             exp.add_step(Step.unzip_exp_dir(exp))
 
+        .. deprecated:: 1.8
+
+            Please define your extra experiment steps manually.
+
         """
         return cls('unzip-exp-dir', call,
                    ['tar', '-xjf', exp.name + '.tar.bz2'],
                    cwd=os.path.dirname(exp.path))
 
     @classmethod
+    @tools.deprecated(_predefined_steps_warning)
     def remove_exp_dir(cls, exp):
         """Return a Step that removes the experiment directory.
 
         ::
 
             exp.add_step(Step.remove_exp_dir(exp))
+
+        .. deprecated:: 1.8
+
+            Please define your extra experiment steps manually.
 
         """
         return cls('remove-exp-dir', shutil.rmtree, exp.path)