+++ tests/whitelist.py
@@ -9,15 +9,16 @@ from downward.experiments.comparerevisions import CompareRevisionsExperiment
 
 from examples import standard_exp
 
+
 lab.experiment.ARGPARSER.epilog
 lab.tools.RawAndDefaultsHelpFormatter._fill_text
 lab.tools.RawAndDefaultsHelpFormatter._get_help_string
 GkiGridEnvironment()
-call.Call(['ls'], stdout='/dev/null')
+call.Call
 log.redirects
 log.driver_log
 log.driver_err
-log.print_(open('/dev/null', 'w'), 'Test')
+log.print_
 log.save_returncode
 lab.steps.Step.unzip_exp_dir
 lab.steps.Step.remove_exp_dir
@@ -40,5 +41,5 @@ lab.steps.Step.remove_exp_dir
 ]
 
 CompareRevisionsExperiment
-standard_exp.get_exp('gripper', [('blind', ['--search', 'astar(blind())'])])
+standard_exp.get_exp
 standard_exp.EXPS